+++ R/check_errors.r
@@ -82,8 +82,11 @@ check_errors <- function(data,
     stop("xlab must be of length 1 or the same length as ci_column.")
 
   # Check tick_digits
-  if(!is.numeric(ticks_digits) || !length(ticks_digits) %in% c(1, length(ci_column)))
-      stop("ticks_digits must be numeric of length 1 or same length as ci_column.")
+  if(!length(ticks_digits) %in% c(1, length(ci_column)))
+      stop("ticks_digits must be length of 1 or same length as ci_column.")
+  
+  if(!is.numeric(unlist(ticks_digits)))
+      stop("ticks_digits must be numeric.")
 
   # If only one CI column
   if(length(ci_column) == 1){