+++ R/forest.R
@@ -12,8 +12,7 @@
 #' @param lower Lower bound of the confidence interval, same as \code{est}.
 #' @param upper Upper bound of the confidence interval, same as \code{est}.
 #' @param sizes Size of the point estimation box, can be a unit, vector or a list.
-#' If the value is not unique, this will first calculate square root of the
-#' reciprocal of size, then devide by overall maximum calculated value.
+#' Values will be used as it is, no transformation will be applied.
 #' @param ref_line X-axis coordinates of zero line, default is 1. Provide an atomic
 #'  vector if different reference line for each \code{ci_column} is desired.
 #' @param vert_line Numerical vector, add additional vertical line at given value.
@@ -113,15 +112,18 @@ forest <- function(data,
   if(!is.null(ticks_at) && !inherits(ticks_at, "list"))
     ticks_at <- rep(list(ticks_at), length(ci_column))
 
-  # ticks digits to accomodate ticks_at
-  if(ticks_digits == 1L & !is.null(ticks_at)){
-    if(is.list(ticks_at))
-      ticks_digits <- sapply(ticks_at, function(x){
-        max(nchar(gsub(".*\\.|^[^.]+$", "", as.character(x))))
-      })
-    else
-      ticks_digits <- max(nchar(gsub(".*\\.|^[^.]+$", "", as.character(ticks_digits))))
+  # ticks digits to accommodate ticks_at
+  if(length(ticks_digits) == 1 & !is.list(ticks_digits)){
+    if(ticks_digits == 1L & !is.null(ticks_at)){
+      if(is.list(ticks_at))
+        ticks_digits <- sapply(ticks_at, function(x){
+          max(nchar(gsub(".*\\.|^[^.]+$", "", as.character(x))))
+        })
+      else
+        ticks_digits <- max(nchar(gsub(".*\\.|^[^.]+$", "", as.character(ticks_digits))))
+    }
   }
+  
 
   if(length(ci_column) != length(ticks_digits))
     ticks_digits <- rep(ticks_digits, length(ci_column))
@@ -190,20 +192,21 @@ forest <- function(data,
     is_summary <- rep(FALSE, nrow(data))
   }
 
-  if(length(unique(stats::na.omit(unlist(sizes)))) != 1){
-    # Get the maximum reciprocal of size
-    max_sizes <- sapply(sizes, function(x){
-      x <- sqrt(1/x)
-      max(x[!is_summary], na.rm = TRUE)
-    }, USE.NAMES = FALSE)
-
-    sizes <- lapply(sizes, function(x){
-      wi <- sqrt(1/x)
-      wi <- wi/max(max_sizes, na.rm = TRUE)
-      wi[is_summary] <- 1/length(max_sizes)
-      return(wi)
-    })
-  }
+  # Transform sizes if not unique and transformation is required.
+  # if(length(unique(stats::na.omit(unlist(sizes)))) != 1 & sizes_trans & group_num == 1){
+  #   # Get the maximum reciprocal of size
+  #   max_sizes <- sapply(sizes, function(x){
+  #     x <- sqrt(x)
+  #     max(x[!is_summary], na.rm = TRUE)
+  #   }, USE.NAMES = FALSE)
+  #
+  #   sizes <- lapply(sizes, function(x){
+  #     wi <- sqrt(x)
+  #     wi <- wi/max(max_sizes, na.rm = TRUE)
+  #     wi[is_summary] <- 1/length(max_sizes)
+  #     return(wi)
+  #   })
+  # }
 
   # Positions of values in ci_column
   gp_list <- rep_len(1:(length(lower)/group_num), length(lower))
@@ -266,6 +269,9 @@ forest <- function(data,
                     theme = theme$tab_theme,
                     rows = NULL)
   }
+
+  # Do not clip text
+  gt$layout$clip <- "off"
 
   # Column index
   col_indx <- rep_len(1:length(ci_column), length(ci_col_list))
@@ -341,7 +347,7 @@ forest <- function(data,
   x_axis <- lapply(seq_along(xlim), function(i){
     make_xaxis(at = ticks_at[[i]],
                gp = theme$xaxis,
-               ticks_digits = ticks_digits[i],
+               ticks_digits = ticks_digits[[i]],
                x0 = ref_line[i],
                xlim = xlim[[i]],
                xlab = xlab[i],