+++ tests/testthat/test-forest.R
@@ -205,8 +205,8 @@ test_that("Multiple column and Multi parameters", {
               x_trans = c("log", "none"),
               arrow_lab = list(c("L1", "R1"), c("L2", "R2")),
               xlim = list(c(0, 3), c(-1, 3)),
-              ticks_at = list(c(0.1, 0.5, 1, 2.5), c(-1, 0, 2)),
-              ticks_digits = 1,
+              ticks_at = list(c(0.1, 0.5, 1, 2.5), c(-1.0, 0, 1.5, 2.0)),
+              ticks_digits = list(1, 1L),
               xlab = c("OR", "Beta"),
               nudge_y = 0.2,
               theme = tm)
@@ -294,7 +294,7 @@ test_that("forestplot check ERRORS", {
 # Check arrow
 test_that("check arrow", {
 
-  dt <- dt[1:9, ]
+  dt <- dt[1:10, ]
   tm <- forest_theme(arrow_cex = .5,
                      arrow_label_just = "end",
                      xaxis_cex = .5,
@@ -335,7 +335,7 @@ test_that("check arrow", {
 
 test_that("x-scale trans", {
 
-  dt <- dt[1:9, ]
+  dt <- dt[1:10, ]
   dt$hi <- dt$hi * 3
 
   dt$hi[9] <- 8
@@ -385,7 +385,7 @@ test_that("x-scale trans", {
 
 test_that("x-scale trans", {
 
-  dt <- dt[1:9, 1:6]
+  dt <- dt[1:10, 1:6]
 
   dt$se <- seq.int(0.2, 2, length.out = length(dt$est))